requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
python-dateutil==2.9.0.post0
pydantic==2.9.0
PyYAML==6.0.1
rich==13.7.1
jinja2==3.1.4
tenacity==8.4.2
tqdm==4.66.5
openai>=1.40.0
Flask==3.0.3
psycopg2-binary==2.9.9
